clean up and changed login for sql inejction
