+++ router/app.js
@@ -1,12 +1,13 @@
 import express from 'express'
 import admin from './admin_api.js';
-import { db } from '../database/database.js';
-import jwt from 'jsonwebtoken';
-
+import login from './login.js';
+import { authenticateToken  } from '../middleware/auth.js';
 
 const router = express.Router();
 
 router.use("/api/admin", admin)
+
+router.use("/", login)
 
 router.get('/api/ping', authenticateToken,  async(req, res) => {
     res.json({
@@ -15,73 +16,4 @@ router.get('/api/ping', authenticateToken,  async(req, res) => {
     });
 });
 
-// Benutzerlogin
-router.post('/login', async (req, res) => {
-    const { username, password } = req.body;
-    console.log('Body:', req.body);
-
-    const ACCESS_TOKEN_SECRET = "LAbyZljkfHH2fCOwcoUi8LKsGKCgcDax2b7ghYUVkATABsB4WFt4M8WjVQgjjoGP"
-
-    try {
-        const query = `SELECT * FROM users WHERE username = '${username}' AND password = '${password}'`;
-        //const query = `SELECT * FROM users WHERE username=? AND password=?`;
-        console.log('SQL-query:', query);
-    
-        db.all(query, (err, rows) => {
-            if (err) {
-                console.error('Database query error:', err.message);
-                return res.status(500).json({
-                    status: '500',
-                    response: 'Internal server error.',
-                    error: err.message
-                });
-            }
-
-            if (rows.length > 0) {
-                const user = rows[0];
-
-                const token = jwt.sign(
-                    { id: user.id, username: user.username, role: user.role },
-                    ACCESS_TOKEN_SECRET,
-                    { expiresIn: '1h' } 
-                );
-
-                return res.status(200).json({ 
-                    token: token, 
-                    username: user.username,
-                    role: user.role,
-                    response: 'Login erfolgreich!' });
-            } else {
-                return res.status(401).json({ response: 'Benutzer nicht gefunden oder falsches Passwort' });
-            }
-        });
-    } catch (error) {
-        console.error(error);
-        return res.status(500).json({ response: 'Internal server error' });
-    }
-});
-
-function authenticateToken(req, res, next) {
-    const token = req.headers['authorization'];
-    if (!token) return res.status(401).json({ response: 'Kein Token' });
-
-    jwt.verify(token, ACCESS_TOKEN_SECRET, (err, user) => {
-        if (err) return res.status(403).json({ response: 'Ungültiges Token' });
-        req.user = user;
-        next();
-    });
-}
-
-function authorizeRole(role) {
-    return (req, res, next) => {
-        if (req.user && req.user.role === role) {
-            next();
-        } else {
-            res.status(403).json({ response: "Zugriff verweigert."})
-        }
-    };
-}
-
-export { authorizeRole };
-export { authenticateToken };
 export default router;
